fixed coorfinates to double

+++ SOS_BE/SOS.Core/Domain/CoordinatesSpan.cs
@@ -9,9 +9,9 @@ namespace SOS.Core.Domain
     public class CoordinatesSpan
     {
         public int Id { get; set; }
-        public int LatitudeStart { get; set; }
-        public int LongitudeStart { get; set; }
-        public int LatitudeEnd { get; set; }
-        public int LongitudeEnd { get; set;}
+        public double LatitudeStart { get; set; }
+        public double LongitudeStart { get; set; }
+        public double LatitudeEnd { get; set; }
+        public double LongitudeEnd { get; set;}
     }
 }
